duplication compiler error fixed
:D
