+++ src/Makefile
@@ -1,6 +1,6 @@
 CC = g++
 CFLAGS = -I.
-OBJS = main.o Color.o 
+OBJS = main.o Vect.o Color.o Light.o 
 
 all: raytracer
 
@@ -11,9 +11,14 @@ raytracer: $(OBJS)
 main.o: main.cpp
 	$(CC) -c main.cpp $(CFLAGS)
 
+Vect.o: Vect.cpp
+	$(CC) -c Vect.cpp $(CFLAGS)
+
 Color.o: Color.cpp
 	$(CC) -c Color.cpp $(CFLAGS)
 
+Light.o: Light.cpp
+	$(CC) -c Light.cpp $(CFLAGS)
 
 clean:
 	rm -f raytracer $(OBJS)
